requirements.txt
# Core dependencies
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Data manipulation
pandas==2.1.4
numpy==1.26.2
polars==0.20.2

# Exchange connectivity
ccxt==4.2.0
websockets==12.0

# Async & HTTP
asyncio==3.4.3
aiohttp==3.9.1
httpx==0.25.2

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b

# Machine Learning & NLP
torch==2.1.2
transformers==4.36.2
scikit-learn==1.3.2
lightgbm==4.2.0
xgboost==2.0.3
spacy==3.7.2

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Feature Engineering & RAG
llama-index==0.9.45
langchain==0.1.0
keybert==0.8.4

# MLOps
mlflow==2.9.2
feast==0.35.0

# Database & Storage
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23
timescale-vector==0.0.1

# Streaming
kafka-python==2.0.2

# Backtesting
vectorbt==0.26.1
backtrader==1.9.78.123

# Workflow orchestration
prefect==2.14.13

# Monitoring & Observability
prometheus-client==0.19.0
grafana-client==4.0.0

# Utilities
requests==2.31.0
PyYAML==6.0.1
click==8.1.7
rich==13.7.0
loguru==0.7.2

# Security
cryptography==41.0.7
hvac==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
